Crate conjure_cp_core

Modules

ast

context

error

Top-level error types for Conjure-Oxide.

parse

representation

rule_engine

solver

A high-level API for interacting with constraints solvers.

stats

Macros

bug

Triggers a panic with a detailed bug report message, while ensuring the panic is ignored in coverage reports.

domain_int

Creates a Domain::Int.

error

into_matrix

Creates a new matrix AbstractLiteral from some Vec, optionally with some index domain.

into_matrix_expr

Creates a new matrix as an Expression from a (Rust) vector, optionally with some index domain.

matrix

Creates a new matrix AbstractLiteral optionally with some index domain.

matrix_expr

Creates a new matrix as an Expression, optionally with some index domain.

range

Creates a Range.

register_representation

throw_error