conjure_cp_core::ast::abstract_comprehension

Struct AbstractComprehensionBuilder

pub struct AbstractComprehensionBuilder {
    pub qualifiers: Vec<Qualifier>,
    pub return_expr_symbols: Rc<RefCell<SymbolTable>>,
    pub generator_symbols: Rc<RefCell<SymbolTable>>,
}

Fields
qualifiers: Vec<Qualifier>return_expr_symbols: Rc<RefCell<SymbolTable>>

The symbol table used in the return expression.

Variables from generator expressions are “given” in the context of the return expression. That is, they are constants which are different for each expansion of the comprehension.

generator_symbols: Rc<RefCell<SymbolTable>>

The scope for variables in generator expressions.

Variables declared in generator expressions are decision variables in their original context, since they do not have a constant value.

Implementations

impl AbstractComprehensionBuilder

pub fn new(symbols: &Rc<RefCell<SymbolTable>>) -> Self

Creates an AbstractComprehensionBuilder with:

An inner scope which inherits from the given symbol table

An empty list of qualifiers

Changes to the inner scope do not affect the given symbol table.

The return expression is passed when finalizing the comprehension, in [with_return_value].

pub fn return_expr_symbols(&self) -> Rc<RefCell<SymbolTable>>

pub fn generator_symbols(&self) -> Rc<RefCell<SymbolTable>>

pub fn new_domain_generator(&mut self, domain: DomainPtr) -> DeclarationPtr

pub fn new_expression_generator(self, expr: Expression, name: Name) -> Self

Creates a new expression generator with the given expression and variable name.

The variable “takes from” the expression, that is, it can be any element in the expression.

E.g. in [ x | x <- some_set ], x can be any element of some_set.

pub fn add_condition(&mut self, condition: Expression)

See crate::ast::comprehension::ComprehensionBuilder::guard

pub fn new_letting(&mut self, expression: Expression) -> DeclarationPtr

pub fn with_return_value(self, expression: Expression) -> AbstractComprehension