conjure_cp_core::ast

Enum Domain

pub enum Domain {
    Bool,
    Int(Vec<Range<i32>>),
    Empty(ReturnType),
    Reference(Name),
    Set(SetAttr, Box<Domain>),
    Matrix(Box<Domain>, Vec<Domain>),
    Tuple(Vec<Domain>),
    Record(Vec<RecordEntry>),
}

Variants

Bool

Int(Vec<Range<i32>>)

An integer domain.

If multiple ranges are inside the domain, the values in the domain are the union of these ranges.

If no ranges are given, the int domain is considered unconstrained, and can take any integer value.

Empty(ReturnType)

An empty domain of the given type.

Reference(Name)

Set(SetAttr, Box<Domain>)

Matrix(Box<Domain>, Vec<Domain>)

A n-dimensional matrix with a value domain and n-index domains

Tuple(Vec<Domain>)

Record(Vec<RecordEntry>)

Implementations

impl Domain

pub fn contains(&self, lit: &Literal) -> Result<bool, DomainOpError>

Returns true if lit is a member of the domain.

Errors

DomainOpError::InputContainsReference if the input domain is a reference or contains a reference, meaning its members cannot be determined.

pub fn values_i32(&self) -> Result<Vec<i32>, DomainOpError>

Returns a list of all possible values in an integer domain.

Errors

DomainOpError::InputNotInteger if the domain is not an integer domain.

DomainOpError::InputUnbounded if the domain is unbounded.

pub fn from_slice_i32(elements: &[i32]) -> Domain

Creates an Domain::Int containing the given integers.

Domain::from_set_i32 should be used instead where possible, as it is cheaper (it does not need to sort its input).

Examples

use conjure_cp_core::ast::{Domain,Range};
use conjure_cp_core::{domain_int,range};

let elements = vec![1,2,3,4,5];

let domain = Domain::from_slice_i32(&elements);

assert_eq!(domain,domain_int!(1..5));

use conjure_cp_core::ast::{Domain,Range};
use conjure_cp_core::{domain_int,range};

let elements = vec![1,2,4,5,7,8,9,10];

let domain = Domain::from_slice_i32(&elements);

assert_eq!(domain,domain_int!(1..2,4..5,7..10));

use conjure_cp_core::ast::{Domain,Range,ReturnType};

let elements = vec![];

let domain = Domain::from_slice_i32(&elements);

assert!(matches!(domain,Domain::Empty(ReturnType::Int)))

pub fn from_set_i32(elements: &BTreeSet<i32>) -> Domain

Creates an Domain::Int containing the given integers.

Examples

use conjure_cp_core::ast::{Domain,Range};
use conjure_cp_core::{domain_int,range};
use std::collections::BTreeSet;

let elements = BTreeSet::from([1,2,3,4,5]);

let domain = Domain::from_set_i32(&elements);

assert_eq!(domain,domain_int!(1..5));

use conjure_cp_core::ast::{Domain,Range};
use conjure_cp_core::{domain_int,range};
use std::collections::BTreeSet;

let elements = BTreeSet::from([1,2,4,5,7,8,9,10]);

let domain = Domain::from_set_i32(&elements);

assert_eq!(domain,domain_int!(1..2,4..5,7..10));

use conjure_cp_core::ast::{Domain,Range,ReturnType};
use std::collections::BTreeSet;

let elements = BTreeSet::from([]);

let domain = Domain::from_set_i32(&elements);

assert!(matches!(domain,Domain::Empty(ReturnType::Int)))

pub fn from_literal_vec(literals: Vec<Literal>) -> Result<Domain, DomainOpError>

For a vector of literals, creates a domain that contains all the elements.

The literals must all be of the same type.

For abstract literals, this method merges the element domains of the literals, but not the index domains. Thus, for fixed-sized abstract literals (matrices, tuples, records, etc.), all literals in the vector must also have the same size / index domain:

Matrices: all literals must have the same index domain.

Tuples: all literals must have the same number of elements.

Records: all literals must have the same fields.

Errors

DomainOpError::InputWrongType if the input literals are of a different type to each-other, as described above.

Examples

use conjure_cp_core::ast::{Domain,Range,Literal,ReturnType};

let domain = Domain::from_literal_vec(vec![]);
assert_eq!(domain,Ok(Domain::Empty(ReturnType::Unknown)));

use conjure_cp_core::ast::{Domain,Range,Literal, AbstractLiteral};
use conjure_cp_core::{domain_int, range, matrix};

// `[1,2;int(2..3)], [4,5; int(2..3)]` has domain
// `matrix indexed by [int(2..3)] of int(1..2,4..5)`

let matrix_1 = Literal::AbstractLiteral(matrix![Literal::Int(1),Literal::Int(2);domain_int!(2..3)]);
let matrix_2 = Literal::AbstractLiteral(matrix![Literal::Int(4),Literal::Int(5);domain_int!(2..3)]);

let domain = Domain::from_literal_vec(vec![matrix_1,matrix_2]);

let expected_domain = Ok(Domain::Matrix(
    Box::new(domain_int!(1..2,4..5)),vec![domain_int!(2..3)]));

assert_eq!(domain,expected_domain);

use conjure_cp_core::ast::{Domain,Range,Literal, AbstractLiteral,DomainOpError};
use conjure_cp_core::{domain_int, range, matrix};

// `[1,2;int(2..3)], [4,5; int(1..2)]` cannot be combined
// `matrix indexed by [int(2..3)] of int(1..2,4..5)`

let matrix_1 = Literal::AbstractLiteral(matrix![Literal::Int(1),Literal::Int(2);domain_int!(2..3)]);
let matrix_2 = Literal::AbstractLiteral(matrix![Literal::Int(4),Literal::Int(5);domain_int!(1..2)]);

let domain = Domain::from_literal_vec(vec![matrix_1,matrix_2]);

assert_eq!(domain,Err(DomainOpError::InputWrongType));

use conjure_cp_core::ast::{Domain,Range,Literal, AbstractLiteral};
use conjure_cp_core::{domain_int,range, matrix};

// `[[1,2; int(1..2)];int(2)], [[4,5; int(1..2)]; int(2)]` has domain
// `matrix indexed by [int(2),int(1..2)] of int(1..2,4..5)`

let matrix_1 = Literal::AbstractLiteral(matrix![Literal::AbstractLiteral(matrix![Literal::Int(1),Literal::Int(2);domain_int!(1..2)]); domain_int!(2)]);
let matrix_2 = Literal::AbstractLiteral(matrix![Literal::AbstractLiteral(matrix![Literal::Int(4),Literal::Int(5);domain_int!(1..2)]); domain_int!(2)]);

let domain = Domain::from_literal_vec(vec![matrix_1,matrix_2]);

let expected_domain = Ok(Domain::Matrix(
    Box::new(domain_int!(1..2,4..5)),
    vec![domain_int!(2),domain_int!(1..2)]));

assert_eq!(domain,expected_domain);

pub fn values(&self) -> Result<Vec<Literal>, DomainOpError>

Gets all the Literal values inside this domain.

Errors

DomainOpError::InputNotInteger if the domain is not an integer domain.

DomainOpError::InputContainsReference if the domain is a reference or contains a reference, meaning that its values cannot be determined.

pub fn length(&self) -> Result<u64, DomainOpError>

Gets the length of this domain.

Errors

DomainOpError::InputUnbounded if the input domain is of infinite size.

DomainOpError::InputContainsReference if the input domain is or contains a domain reference, meaning that its size cannot be determined.

pub fn apply_i32( &self, op: fn(i32, i32) -> Option<i32>, other: &Domain, ) -> Result<Domain, DomainOpError>

Returns the domain that is the result of applying a binary operation to two integer domains.

The given operator may return None if the operation is not defined for its arguments. Undefined values will not be included in the resulting domain.

Errors

DomainOpError::InputUnbounded if either of the input domains are unbounded.

DomainOpError::InputNotInteger if either of the input domains are not integers.

pub fn is_finite(&self) -> Result<bool, DomainOpError>

Returns true if the domain is finite.

Errors

DomainOpError::InputContainsReference if the input domain is or contains a domain reference, meaning that its size cannot be determined.

pub fn resolve(self, symbols: &SymbolTable) -> Domain

Resolves this domain to a ground domain, using the symbol table provided to resolve references.

A domain is ground iff it is not a domain reference, nor contains any domain references.

See also: SymbolTable::resolve_domain.

Panics

If a reference domain in self does not exist in the given symbol table.

pub fn intersect(&self, other: &Domain) -> Result<Domain, DomainOpError>

Calculates the intersection of two domains.

Errors

DomainOpError::InputUnbounded if either of the input domains are unbounded.

DomainOpError::InputWrongType if the input domains are different types, or are not integer or set domains.

pub fn union(&self, other: &Domain) -> Result<Domain, DomainOpError>

Calculates the union of two domains.

Errors

DomainOpError::InputUnbounded if either of the input domains are unbounded.

DomainOpError::InputWrongType if the input domains are different types, or are not integer set, or matrix domains. This is also thrown if the matrix domains that have different index domains.