conjure_cp_core::ast

Type Alias DomainPtr

pub type DomainPtr = Moo<Domain>;

Aliased Type

pub struct DomainPtr { /* private fields */ }

Implementations

impl DomainPtr

pub fn resolve(&self) -> Option<Moo<GroundDomain>>

pub fn union(&self, other: &DomainPtr) -> Result<DomainPtr, DomainOpError>

Convenience method to take Domain::union of the Domains behind two DomainPtrs and wrap the result in a new DomainPtr.

pub fn intersect(&self, other: &DomainPtr) -> Result<DomainPtr, DomainOpError>

Convenience method to take Domain::intersect of the Domains behind two DomainPtrs and wrap the result in a new DomainPtr.