conjure_cp_core

Macro bug

macro_rules! bug {
    ($msg:expr) => { ... };
    ($msg:expr, $($arg:tt)*) => { ... };
}

Triggers a panic with a detailed bug report message, while ensuring the panic is ignored in coverage reports.

This macro is useful in situations where an unreachable code path is hit or when a bug occurs.

Parameters

msg: A string expression describing the cause of the panic or bug.