conjure_cp_core::ast

Enum Expression

pub enum Expression {

    AbstractLiteral(Metadata, AbstractLiteral<Expression>),
    Root(Metadata, Vec<Expression>),
    Bubble(Metadata, Moo<Expression>, Moo<Expression>),
    Comprehension(Metadata, Moo<Comprehension>),
    AbstractComprehension(Metadata, Moo<AbstractComprehension>),
    DominanceRelation(Metadata, Moo<Expression>),
    FromSolution(Metadata, Moo<Atom>),
    Metavar(Metadata, Ustr),
    Atomic(Metadata, Atom),
    UnsafeIndex(Metadata, Moo<Expression>, Vec<Expression>),
    SafeIndex(Metadata, Moo<Expression>, Vec<Expression>),
    UnsafeSlice(Metadata, Moo<Expression>, Vec<Option<Expression>>),
    SafeSlice(Metadata, Moo<Expression>, Vec<Option<Expression>>),
    InDomain(Metadata, Moo<Expression>, DomainPtr),
    ToInt(Metadata, Moo<Expression>),
    Scope(Metadata, Moo<SubModel>),
    Abs(Metadata, Moo<Expression>),
    Sum(Metadata, Moo<Expression>),
    Product(Metadata, Moo<Expression>),
    Min(Metadata, Moo<Expression>),
    Max(Metadata, Moo<Expression>),
    Not(Metadata, Moo<Expression>),
    Or(Metadata, Moo<Expression>),
    And(Metadata, Moo<Expression>),
    Imply(Metadata, Moo<Expression>, Moo<Expression>),
    Iff(Metadata, Moo<Expression>, Moo<Expression>),
    Union(Metadata, Moo<Expression>, Moo<Expression>),
    In(Metadata, Moo<Expression>, Moo<Expression>),
    Intersect(Metadata, Moo<Expression>, Moo<Expression>),
    Supset(Metadata, Moo<Expression>, Moo<Expression>),
    SupsetEq(Metadata, Moo<Expression>, Moo<Expression>),
    Subset(Metadata, Moo<Expression>, Moo<Expression>),
    SubsetEq(Metadata, Moo<Expression>, Moo<Expression>),
    Eq(Metadata, Moo<Expression>, Moo<Expression>),
    Neq(Metadata, Moo<Expression>, Moo<Expression>),
    Geq(Metadata, Moo<Expression>, Moo<Expression>),
    Leq(Metadata, Moo<Expression>, Moo<Expression>),
    Gt(Metadata, Moo<Expression>, Moo<Expression>),
    Lt(Metadata, Moo<Expression>, Moo<Expression>),
    SafeDiv(Metadata, Moo<Expression>, Moo<Expression>),
    UnsafeDiv(Metadata, Moo<Expression>, Moo<Expression>),
    SafeMod(Metadata, Moo<Expression>, Moo<Expression>),
    UnsafeMod(Metadata, Moo<Expression>, Moo<Expression>),
    Neg(Metadata, Moo<Expression>),
    Defined(Metadata, Moo<Expression>),
    Range(Metadata, Moo<Expression>),
    UnsafePow(Metadata, Moo<Expression>, Moo<Expression>),
    SafePow(Metadata, Moo<Expression>, Moo<Expression>),
    Flatten(Metadata, Option<Moo<Expression>>, Moo<Expression>),
    AllDiff(Metadata, Moo<Expression>),
    Minus(Metadata, Moo<Expression>, Moo<Expression>),
    FlatAbsEq(Metadata, Moo<Atom>, Moo<Atom>),
    FlatAllDiff(Metadata, Vec<Atom>),
    FlatSumGeq(Metadata, Vec<Atom>, Atom),
    FlatSumLeq(Metadata, Vec<Atom>, Atom),
    FlatIneq(Metadata, Moo<Atom>, Moo<Atom>, Box<Literal>),
    FlatWatchedLiteral(Metadata, Reference, Literal),
    FlatWeightedSumLeq(Metadata, Vec<Literal>, Vec<Atom>, Moo<Atom>),
    FlatWeightedSumGeq(Metadata, Vec<Literal>, Vec<Atom>, Moo<Atom>),
    FlatMinusEq(Metadata, Moo<Atom>, Moo<Atom>),
    FlatProductEq(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>),
    MinionDivEqUndefZero(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>),
    MinionModuloEqUndefZero(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>),
    MinionPow(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>),
    MinionReify(Metadata, Moo<Expression>, Atom),
    MinionReifyImply(Metadata, Moo<Expression>, Atom),
    MinionWInIntervalSet(Metadata, Atom, Vec<i32>),
    MinionWInSet(Metadata, Atom, Vec<i32>),
    MinionElementOne(Metadata, Vec<Atom>, Moo<Atom>, Moo<Atom>),
    AuxDeclaration(Metadata, Reference, Moo<Expression>),
    SATInt(Metadata, SATIntEncoding, Moo<Expression>, (i32, i32)),
    PairwiseSum(Metadata, Moo<Expression>, Moo<Expression>),
    PairwiseProduct(Metadata, Moo<Expression>, Moo<Expression>),
    Image(Metadata, Moo<Expression>, Moo<Expression>),
    ImageSet(Metadata, Moo<Expression>, Moo<Expression>),
    PreImage(Metadata, Moo<Expression>, Moo<Expression>),
    Inverse(Metadata, Moo<Expression>, Moo<Expression>),
    Restrict(Metadata, Moo<Expression>, Moo<Expression>),
    LexLt(Metadata, Moo<Expression>, Moo<Expression>),
    LexLeq(Metadata, Moo<Expression>, Moo<Expression>),
    LexGt(Metadata, Moo<Expression>, Moo<Expression>),
    LexGeq(Metadata, Moo<Expression>, Moo<Expression>),
    FlatLexLt(Metadata, Vec<Atom>, Vec<Atom>),
    FlatLexLeq(Metadata, Vec<Atom>, Vec<Atom>),

}

Represents different types of expressions used to define rules and constraints in the model.

The Expression enum includes operations, constants, and variable references used to build rules and conditions for the model.

Compatability

SMT

Abs

And

Eq

Geq

Gt

Iff

Imply

Leq

Lt

Max

Min

Neg

Neq

Not

Or

PairwiseProduct

PairwiseSum

Product

SafeDiv

SafeIndex

SafeMod

Sum

ToInt

SAT

And

Not

Or

SATInt

Minion

AuxDeclaration

FlatAbsEq

FlatAllDiff

FlatIneq

FlatMinusEq

FlatProductEq

FlatSumGeq

FlatSumLeq

FlatWatchedLiteral

MinionDivEqUndefZero

MinionElementOne

MinionModuloEqUndefZero

MinionReify

MinionReifyImply

MinionWInIntervalSet

MinionWInSet

JsonInput

Abs

AllDiff

And

Defined

Eq

Geq

Gt

Iff

Image

ImageSet

Imply

In

Intersect

Inverse

Leq

Lt

Max

Min

Minus

Neg

Neq

Not

Or

PreImage

Product

Range

Restrict

Subset

SubsetEq

Sum

Supset

SupsetEq

Union

UnsafeDiv

UnsafeIndex

UnsafeMod

UnsafePow

UnsafeSlice

Variants

AbstractLiteral(Metadata, AbstractLiteral<Expression>)

Root(Metadata, Vec<Expression>)

The top of the model

Bubble(Metadata, Moo<Expression>, Moo<Expression>)

An expression representing A is valid as long as B is true Turns into a conjunction when it reaches a boolean context

Comprehension(Metadata, Moo<Comprehension>)

A comprehension.

The inside of the comprehension opens a new scope.

AbstractComprehension(Metadata, Moo<AbstractComprehension>)

Higher-level abstract comprehension

DominanceRelation(Metadata, Moo<Expression>)

Defines dominance (Solution A is preferred over Solution B)

FromSolution(Metadata, Moo<Atom>)

fromSolution(name) - Used in dominance relation definitions

Metavar(Metadata, Ustr)

Atomic(Metadata, Atom)

UnsafeIndex(Metadata, Moo<Expression>, Vec<Expression>)

A matrix index.

Defined iff the indices are within their respective index domains. Supported by: JsonInput.

SafeIndex(Metadata, Moo<Expression>, Vec<Expression>)

A safe matrix index.

See Expression::UnsafeIndex Supported by: SMT.

UnsafeSlice(Metadata, Moo<Expression>, Vec<Option<Expression>>)

A matrix slice: a[indices].

One of the indicies may be None, representing the dimension of the matrix we want to take a slice of. For example, for some 3d matrix a, a[1,..,2] has the indices Some(1),None,Some(2).

It is assumed that the slice only has one wild-card dimension and thus is 1 dimensional.

Defined iff the defined indices are within their respective index domains. Supported by: JsonInput.

SafeSlice(Metadata, Moo<Expression>, Vec<Option<Expression>>)

A safe matrix slice: a[indices].

See Expression::UnsafeSlice.

InDomain(Metadata, Moo<Expression>, DomainPtr)

inDomain(x,domain) iff x is in the domain domain.

This cannot be constructed from Essence input, nor passed to a solver: this expression is mainly used during the conversion of UnsafeIndex and UnsafeSlice to SafeIndex and SafeSlice respectively.

ToInt(Metadata, Moo<Expression>)

toInt(b) casts boolean expression b to an integer.

If b is false, then toInt(b) == 0

If b is true, then toInt(b) == 1 Supported by: SMT.

Scope(Metadata, Moo<SubModel>)

Abs(Metadata, Moo<Expression>)

|x| - absolute value of x Supported by: JsonInput, SMT.

Sum(Metadata, Moo<Expression>)

sum(<vec_expr>) Supported by: JsonInput, SMT.

Product(Metadata, Moo<Expression>)

a * b * c * ... Supported by: JsonInput, SMT.

Min(Metadata, Moo<Expression>)

min(<vec_expr>) Supported by: JsonInput, SMT.

Max(Metadata, Moo<Expression>)

max(<vec_expr>) Supported by: JsonInput, SMT.

Not(Metadata, Moo<Expression>)

not(a) Supported by: JsonInput, SAT, SMT.

Or(Metadata, Moo<Expression>)

or(<vec_expr>) Supported by: JsonInput, SAT, SMT.

And(Metadata, Moo<Expression>)

and(<vec_expr>) Supported by: JsonInput, SAT, SMT.

Imply(Metadata, Moo<Expression>, Moo<Expression>)

Ensures that a->b (material implication). Supported by: JsonInput, SMT.

Iff(Metadata, Moo<Expression>, Moo<Expression>)

iff(a, b) a <-> b Supported by: JsonInput, SMT.

Union(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

In(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Intersect(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Supset(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

SupsetEq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Subset(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

SubsetEq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Eq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

Neq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

Geq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

Leq(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

Gt(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

Lt(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput, SMT.

SafeDiv(Metadata, Moo<Expression>, Moo<Expression>)

Division after preventing division by zero, usually with a bubble Supported by: SMT.

UnsafeDiv(Metadata, Moo<Expression>, Moo<Expression>)

Division with a possibly undefined value (division by 0) Supported by: JsonInput.

SafeMod(Metadata, Moo<Expression>, Moo<Expression>)

Modulo after preventing mod 0, usually with a bubble Supported by: SMT.

UnsafeMod(Metadata, Moo<Expression>, Moo<Expression>)

Modulo with a possibly undefined value (mod 0) Supported by: JsonInput.

Neg(Metadata, Moo<Expression>)

Negation: -x Supported by: JsonInput, SMT.

Defined(Metadata, Moo<Expression>)

Set of domain values function is defined for Supported by: JsonInput.

Range(Metadata, Moo<Expression>)

Set of codomain values function is defined for Supported by: JsonInput.

UnsafePow(Metadata, Moo<Expression>, Moo<Expression>)

Unsafe powerx**y (possibly undefined)

Defined when (X!=0 \/ Y!=0) /\ Y>=0 Supported by: JsonInput.

SafePow(Metadata, Moo<Expression>, Moo<Expression>)

UnsafePow after preventing undefinedness

Flatten(Metadata, Option<Moo<Expression>>, Moo<Expression>)

Flatten matrix operator flatten(M) or flatten(n, M) where M is a matrix and n is an optional integer argument indicating depth of flattening

AllDiff(Metadata, Moo<Expression>)

allDiff(<vec_expr>) Supported by: JsonInput.

Minus(Metadata, Moo<Expression>, Moo<Expression>)

Binary subtraction operator

This is a parser-level construct, and is immediately normalised to Sum([a,-b]). TODO: make this compatible with Set Difference calculations - need to change return type and domain for this expression and write a set comprehension rule. have already edited minus_to_sum to prevent this from applying to sets Supported by: JsonInput.

FlatAbsEq(Metadata, Moo<Atom>, Moo<Atom>)

Ensures that x=|y| i.e. x is the absolute value of y.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatAllDiff(Metadata, Vec<Atom>)

Ensures that alldiff([a,b,...]).

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatSumGeq(Metadata, Vec<Atom>, Atom)

Ensures that sum(vec) >= x.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatSumLeq(Metadata, Vec<Atom>, Atom)

Ensures that sum(vec) <= x.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatIneq(Metadata, Moo<Atom>, Moo<Atom>, Box<Literal>)

ineq(x,y,k) ensures that x <= y + k.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatWatchedLiteral(Metadata, Reference, Literal)

w-literal(x,k) ensures that x == k, where x is a variable and k a constant.

Low-level Minion constraint.

This is a low-level Minion constraint and you should probably use Eq instead. The main use of w-literal is to convert boolean variables to constraints so that they can be used inside watched-and and watched-or.

See also

Minion documentation

rules::minion::boolean_literal_to_wliteral. Supported by: Minion.

FlatWeightedSumLeq(Metadata, Vec<Literal>, Vec<Atom>, Moo<Atom>)

weightedsumleq(cs,xs,total) ensures that cs.xs <= total, where cs.xs is the scalar dot product of cs and xs.

Low-level Minion constraint.

Represents a weighted sum of the form ax + by + cz + ...

See also

Minion documentation

FlatWeightedSumGeq(Metadata, Vec<Literal>, Vec<Atom>, Moo<Atom>)

weightedsumgeq(cs,xs,total) ensures that cs.xs >= total, where cs.xs is the scalar dot product of cs and xs.

Low-level Minion constraint.

Represents a weighted sum of the form ax + by + cz + ...

See also

Minion documentation

FlatMinusEq(Metadata, Moo<Atom>, Moo<Atom>)

Ensures that x =-y, where x and y are atoms.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

FlatProductEq(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>)

Ensures that x*y=z.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionDivEqUndefZero(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>)

Ensures that floor(x/y)=z. Always true when y=0.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionModuloEqUndefZero(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>)

Ensures that x%y=z. Always true when y=0.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionPow(Metadata, Moo<Atom>, Moo<Atom>, Moo<Atom>)

Ensures that x**y = z.

Low-level Minion constraint.

This constraint is false when y<0 except for 1**y=1 and (-1)**y=z (where z is 1 if y is odd and z is -1 if y is even).

See also

Github comment about pow semantics

Minion documentation

MinionReify(Metadata, Moo<Expression>, Atom)

reify(constraint,r) ensures that r=1 iff constraint is satisfied, where r is a 0/1 variable.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionReifyImply(Metadata, Moo<Expression>, Atom)

reifyimply(constraint,r) ensures that r->constraint, where r is a 0/1 variable. variable.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionWInIntervalSet(Metadata, Atom, Vec<i32>)

w-inintervalset(x, [a1,a2, b1,b2, ]) ensures that the value of x belongs to one of the intervals {a1,,a2}, {b1,,b2} etc.

The list of intervals must be given in numerical order.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionWInSet(Metadata, Atom, Vec<i32>)

w-inset(x, [v1, v2, ]) ensures that the value of x is one of the explicitly given values v1, v2, etc.

This constraint enforces membership in a specific set of discrete values rather than intervals.

The list of values must be given in numerical order.

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

MinionElementOne(Metadata, Vec<Atom>, Moo<Atom>, Moo<Atom>)

element_one(vec, i, e) specifies that vec[i] = e. This implies that i is in the range [1..len(vec)].

Low-level Minion constraint.

See also

Minion documentation Supported by: Minion.

AuxDeclaration(Metadata, Reference, Moo<Expression>)

Declaration of an auxiliary variable.

As with Savile Row, we semantically distinguish this from Eq. Supported by: Minion.

SATInt(Metadata, SATIntEncoding, Moo<Expression>, (i32, i32))

This expression is for encoding ints for the SAT solver, it stores the encoding type, the vector of booleans and the min/max for the int. Supported by: SAT.

PairwiseSum(Metadata, Moo<Expression>, Moo<Expression>)

Addition over a pair of expressions (i.e. a + b) rather than a vec-expr like Expression::Sum. This is for compatibility with backends that do not support addition over vectors. Supported by: SMT.

PairwiseProduct(Metadata, Moo<Expression>, Moo<Expression>)

Multiplication over a pair of expressions (i.e. a * b) rather than a vec-expr like Expression::Product. This is for compatibility with backends that do not support multiplication over vectors. Supported by: SMT.

Image(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

ImageSet(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

PreImage(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Inverse(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

Restrict(Metadata, Moo<Expression>, Moo<Expression>)

Supported by: JsonInput.

LexLt(Metadata, Moo<Expression>, Moo<Expression>)

Lexicographical < between two matrices.

A <lex B iff: A[i] < B[i] for some i /\ (A[j] > B[j] for some j -> i < j) I.e. A must be less than B at some index i, and if it is greater than B at another index j, then j comes after i. I.e. A must be greater than B at the first index where they differ.

E.g. [1, 1] <lex [2, 1] and [1, 1] <lex [1, 2]

LexLeq(Metadata, Moo<Expression>, Moo<Expression>)

Lexicographical <= between two matrices

LexGt(Metadata, Moo<Expression>, Moo<Expression>)

Lexicographical > between two matrices This is a parser-level construct, and is immediately normalised to LexLt(b, a)

LexGeq(Metadata, Moo<Expression>, Moo<Expression>)

Lexicographical >= between two matrices This is a parser-level construct, and is immediately normalised to LexLeq(b, a)

FlatLexLt(Metadata, Vec<Atom>, Vec<Atom>)

Low-level minion constraint. See Expression::LexLt

FlatLexLeq(Metadata, Vec<Atom>, Vec<Atom>)

Low-level minion constraint. See Expression::LexLeq

Implementations

impl Expression

pub fn domain_of(&self) -> Option<DomainPtr>

Returns the possible values of the expression, recursing to leaf expressions

pub fn get_meta(&self) -> Metadata

pub fn set_meta(&self, meta: Metadata)

pub fn is_safe(&self) -> bool

Checks whether this expression is safe.

An expression is unsafe if can be undefined, or if any of its children can be undefined.

Unsafe expressions are (typically) prefixed with Unsafe in our AST, and can be made safe through the use of bubble rules.

pub fn is_clean(&self) -> bool

pub fn set_clean(&mut self, bool_value: bool)

pub fn is_associative_commutative_operator(&self) -> bool

True if the expression is an associative and commutative operator

pub fn is_matrix_literal(&self) -> bool

True if the expression is a matrix literal.

This is true for both forms of matrix literals: those with elements of type Literal and Expression.

pub fn identical_atom_to(&self, other: &Expression) -> bool

True iff self and other are both atomic and identical.

This method is useful to cheaply check equivalence. Assuming CSE is enabled, any unifiable expressions will be rewritten to a common variable. This is much cheaper than checking the entire subtrees of self and other.

pub fn unwrap_list(&self) -> Option<Vec<Expression>>

If the expression is a list, returns a copied vector of the inner expressions.

A list is any a matrix with the domain int(1..). This includes matrix literals without any explicitly specified domain.

pub fn unwrap_matrix_unchecked(self) -> Option<(Vec<Expression>, DomainPtr)>

If the expression is a matrix, gets it elements and index domain.

Consider using the safer Expression::unwrap_list instead.

It is generally undefined to edit the length of a matrix unless it is a list (as defined by Expression::unwrap_list). Users of this function should ensure that, if the matrix is reconstructed, the index domain and the number of elements in the matrix remain the same.

pub fn extend_root(self, exprs: Vec<Expression>) -> Expression

For a Root expression, extends the inner vec with the given vec.

Panics

Panics if the expression is not Root.

pub fn into_literal(self) -> Option<Literal>

Converts the expression to a literal, if possible.

pub fn to_ac_operator_kind(&self) -> Option<ACOperatorKind>

If this expression is an associative-commutative operator, return its ACOperatorKind.

pub fn universe_categories(&self) -> HashSet<Category>

Returns the categories of all sub-expressions of self.